@@ -7,7 +7,7 @@ name = "mitosis"
 dynamic = ["version"]
 description = "Reproduce Machine Learning experiments easily"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.10"
 license = {text = "MIT"}
 keywords = ["Machine Learning", "Science", "Mathematics", "Experiments"]
 authors = [
@@ -34,8 +34,11 @@ classifiers = [
   "Topic :: Text Processing :: Markup",
 ]
 dependencies = [
+  "dill",
   "GitPython",
+  "importlib_metadata",
   "ipykernel",
+  "matplotlib",
   "nbconvert",
   "nbclient",
   "nbformat",
@@ -53,13 +56,19 @@ dev = [
   "flake8-comprehensions>=3.1.0",
   "black",
   "coverage",
+  "mypy",
+  "pandas-stubs",
   "pre-commit",
+  "pytest-lazy-fixture",
   "sphinx",
   "codecov"
 ]
 
 [project.urls]
-homepage = "https://github.com/Jake-Stevens-Haas/mitosis"
+homepage = "https://github.com/Jacob-Stevens-Haas/mitosis"
+
+[project.scripts]
+mitosis="mitosis.__main__:main"
 
 [tool.setuptools_scm]
 
@@ -82,5 +91,8 @@ extend-exclude = '''
 filterwarnings = [
   "ignore::RuntimeWarning",
   "ignore::UserWarning",]
-addopts = '-m "not slow"'
-markers = ["slow"]
+addopts = '-m "not clean"'
+markers = ["clean"]
+
+[tool.codespell]
+ignore-words-list = "raison"
